refactor: move class
